#367 Fixes double slabs drops
