+++ src/main/java/cn/nukkit/block/BlockDoubleSlabStone.java
@@ -3,7 +3,6 @@ package cn.nukkit.block;
 import cn.nukkit.api.PowerNukkitDifference;
 import cn.nukkit.blockproperty.BlockProperties;
 import cn.nukkit.blockproperty.value.StoneSlab1Type;
-import cn.nukkit.item.Item;
 import cn.nukkit.item.ItemTool;
 import cn.nukkit.utils.BlockColor;
 
@@ -63,8 +62,8 @@ public class BlockDoubleSlabStone extends BlockDoubleSlabBase {
     }
 
     @Override
-    public Item toItem() {
-        return getCurrentState().forItem().withBlockId(BlockID.STONE_SLAB).asItemBlock();
+    public int getSingleSlabId() {
+        return STONE_SLAB;
     }
     
     public void setSlabType(StoneSlab1Type type) {
@@ -74,17 +73,6 @@ public class BlockDoubleSlabStone extends BlockDoubleSlabBase {
     @Override
     public String getSlabName() {
         return getSlabType().getEnglishName();
-    }
-
-    @Override
-    public Item[] getDrops(Item item) {
-        if (item.isPickaxe() && item.getTier() >= ItemTool.TIER_WOODEN) {
-            Item slab = toItem();
-            slab.setCount(2);
-            return new Item[]{ slab };
-        } else {
-            return new Item[0];
-        }
     }
 
     @Override
